+++ src/store/index.js
@@ -22,8 +22,9 @@ export default new Vuex.Store({
     bfeState: false,
     otherState: false,
     fwwState: false,
+    noaaState: false,
     currentZoomState: 4,
-    selectedTimePeriodState: "&period=P7D"
+    selectedTimePeriodState: "&period=P7D",
   },
   mutations: {
     toggleDrawerState(state, data) {
@@ -46,6 +47,9 @@ export default new Vuex.Store({
     },
     getFwwState(state, fwwPicked) {
       state.fwwState = fwwPicked;
+    },
+    getNoaaState(state, noaaPicked) {
+      state.noaaState = noaaPicked;
     },
     getBankState(state, bankPicked) {
       state.bankState = bankPicked;
@@ -86,6 +90,7 @@ export default new Vuex.Store({
     nfhlState: (state) => state.nfhlState,
     allRPState: (state) => state.allRPState,
     fwwState: (state) => state.fwwState,
+    noaaState: (state) => state.noaaState,
     currentZoomState: (state) => state.currentZoomState,
     selectedTimePeriodState: (state) => state.selectedTimePeriodState,
   },