+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Add tooltips for active flooding
+- Added NOAA Tides and Current Stations supporting layer
 
 ### Changed